backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
colorama==0.4.6
distlib==0.3.7
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==25.6.0
filelock==3.13.1
iniconfig==2.0.0
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
pipenv==2023.11.17
platformdirs==4.0.0
pluggy==1.5.0
psycopg2==2.9.9
pyproject-api==1.6.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
setuptools==75.2.0
six==1.16.0
sqlparse==0.4.4
tox==4.11.4
tzdata==2023.3
virtualenv==20.24.7
black==23.9.1
